Convert type of pixelHeight/pixelWidth to platform-independent int64_t.

+++ include/exiv2/bmffimage.hpp
@@ -134,8 +134,8 @@ namespace Exiv2
         //! @name Accessors
         //@{
         std::string mimeType() const override /* override */;
-        int pixelWidth() const override;
-        int pixelHeight() const override;
+        int64_t pixelWidth() const override;
+        int64_t pixelHeight() const override;
         //@}
         
         Exiv2::ByteOrder endian_{Exiv2::bigEndian};
