+++ include/exiv2/mrwimage.hpp
@@ -95,8 +95,8 @@ namespace Exiv2 {
         //! @name Accessors
         //@{
         std::string mimeType() const override;
-        int pixelWidth() const override;
-        int pixelHeight() const override;
+        int64_t pixelWidth() const override;
+        int64_t pixelHeight() const override;
         //@}
     }; // class MrwImage
