+++ src/mrwimage.cpp
@@ -49,7 +49,7 @@ namespace Exiv2 {
         return "image/x-minolta-mrw";
     }
 
-    int MrwImage::pixelWidth() const
+    int64_t MrwImage::pixelWidth() const
     {
         auto imageWidth = exifData_.findKey(Exiv2::ExifKey("Exif.Image.ImageWidth"));
         if (imageWidth != exifData_.end() && imageWidth->count() > 0) {
@@ -58,7 +58,7 @@ namespace Exiv2 {
         return 0;
     }
 
-    int MrwImage::pixelHeight() const
+    int64_t MrwImage::pixelHeight() const
     {
         auto imageHeight = exifData_.findKey(Exiv2::ExifKey("Exif.Image.ImageLength"));
         if (imageHeight != exifData_.end() && imageHeight->count() > 0) {
